requirements.txt
Babel>=0.9.6
eventlet>=0.13.0
iso8601>=0.1.8
oslo.config>=1.2.0
pbr>=0.5.21,<1.0
pecan>=0.2.0
pycrypto>=2.6
six>=1.4.1
sphinxcontrib-pecanwsme>=0.5
SQLAlchemy>=0.7.8,<=0.7.99
sqlalchemy-migrate>=0.7.2
WSME>=0.5b6
